A small expedition of program
For the kancollect, the expedition cost time to do. So, the program is a timer, and you can set the time for expedition count the time when the expedition come back.

The program is first version. If you use that it occurs some error, you can via e-mail on my github profile.